Add create-photo.js.
