+++ js/main.js
@@ -1,3 +1,3 @@
-import {createPhotos} from './data.js';
+import {createPhotos} from './create-photos.js';
 
 createPhotos();
